fix: address PR review comments
- Fix bug in textMeasurement.ts where word counting for ellipsis was incorrect
  - Changed from using indexOf() to counting actual words used
- Extract duplicated PDF text utility functions to shared module
  - Created utils/pdf-text-utils.ts for splitTextIntoLines and calculateShrinkToFitFontSize
  - Updated both pages/api/generate.ts and lib/pdf-generator.ts to import from shared module
